+++ src/use-async-combine-race.js
@@ -1,23 +1,27 @@
-import { useRef } from 'react';
+import { useEffect, useRef } from 'react';
 
 import { useAsyncTask } from './use-async-task';
 
 export const useAsyncCombineRace = (...asyncTasks) => {
   const callback = useRef(null);
-  if (callback.current) {
-    callback.current(asyncTasks);
-  }
+  useEffect(() => {
+    if (callback.current) {
+      callback.current(asyncTasks);
+    }
+  });
   const task = useAsyncTask(async (abortController) => {
     abortController.signal.addEventListener('abort', () => {
       asyncTasks.forEach((asyncTask) => {
-        asyncTask.abort();
+        if (asyncTask.abort) {
+          asyncTask.abort();
+        }
       });
     });
     const stopOthers = (tasks) => {
       const index = tasks.findIndex(({ pending }) => !pending);
       if (index >= 0) {
         tasks.forEach((asyncTask, i) => {
-          if (i !== index) {
+          if (i !== index && asyncTask.abort) {
             asyncTask.abort();
           }
         });
@@ -25,9 +29,16 @@ export const useAsyncCombineRace = (...asyncTasks) => {
     };
     callback.current = stopOthers;
     asyncTasks.forEach((asyncTask) => {
+      if (!asyncTask.start) throw new Error('no asyncTask.start');
       asyncTask.start();
     });
-  }, asyncTasks.map(({ taskId }) => taskId));
+  }, asyncTasks.map(({ start }) => start));
+  useEffect(() => {
+    const cleanup = () => {
+      callback.current = null;
+    };
+    return cleanup;
+  }, asyncTasks.map(({ start }) => start));
   return {
     ...task,
     pending: asyncTasks.some(({ pending }) => pending),