+++ src/use-async-combine-seq.js
@@ -1,16 +1,20 @@
-import { useRef } from 'react';
+import { useEffect, useRef } from 'react';
 
 import { useAsyncTask } from './use-async-task';
 
 export const useAsyncCombineSeq = (...asyncTasks) => {
   const callback = useRef(null);
-  if (callback.current) {
-    callback.current(asyncTasks);
-  }
+  useEffect(() => {
+    if (callback.current) {
+      callback.current(asyncTasks);
+    }
+  });
   const task = useAsyncTask(async (abortController) => {
     abortController.signal.addEventListener('abort', () => {
       asyncTasks.forEach((asyncTask) => {
-        asyncTask.abort();
+        if (asyncTask.abort) {
+          asyncTask.abort();
+        }
       });
     });
     const startNext = (tasks) => {
@@ -18,12 +22,20 @@ export const useAsyncCombineSeq = (...asyncTasks) => {
       const prevTask = tasks[index - 1];
       const nextTask = tasks[index];
       if (nextTask && prevTask && !prevTask.pending && !prevTask.error) {
+        if (!nextTask.start) throw new Error('no asyncTask.start');
         nextTask.start();
       }
     };
     callback.current = startNext;
+    if (!asyncTasks[0].start) throw new Error('no asyncTask.start');
     asyncTasks[0].start();
-  }, asyncTasks.map(({ taskId }) => taskId));
+  }, asyncTasks.map(({ start }) => start));
+  useEffect(() => {
+    const cleanup = () => {
+      callback.current = null;
+    };
+    return cleanup;
+  }, asyncTasks.map(({ start }) => start));
   return {
     ...task,
     pending: asyncTasks.some(({ pending }) => pending),