+++ src/use-async-run.js
@@ -1,17 +1,21 @@
 import { useEffect } from 'react';
 
 export const useAsyncRun = (asyncTask) => {
+  const start = asyncTask && asyncTask.start;
+  const abort = asyncTask && asyncTask.abort;
   useEffect(() => {
-    if (asyncTask) {
-      asyncTask.start();
+    if (start) {
+      start();
     }
+  }, [start]);
+  useEffect(() => {
     const cleanup = () => {
-      if (asyncTask) {
-        asyncTask.abort();
+      if (abort) {
+        abort();
       }
     };
     return cleanup;
-  }, [asyncTask && asyncTask.taskId]);
+  }, [abort]);
 };
 
 export default useAsyncRun;