+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Rename "inputs" to "deps"
+- Improve the implementation with useReducer
+- Fix the useRef in useAsyncCombineSeq/Race
 
 ## [1.3.0] - 2019-03-04
 ### Changed